Tread lightly, she is near

+++ 0x15-file_io/0-read_textfile.c
@@ -37,8 +37,8 @@ ssize_t read_textfile(const char *filename, size_t letters)
 
 	buffer[bytesRead] = '\0';
 
-	bytesWritten = write(STDOUT_FILENO, buffer, bytesRead);
-	if (bytesWritten == -1 || (size_t)bytesWritten != bytesRead)
+	bytesWritten = write(STDOUT_FILENO, buffer, (size_t)bytesRead);
+	if (bytesWritten == -1 || (size_t)bytesWritten != (size_t)bytesRead)
 	{
 		fclose(file);
 		free(buffer);
